This PR introduces some changes to the pruning process. The changes don't address witness pruning or adapting the delta parameters, but it implements some changes to the `prune_alpha!` function and adds in a method to prune strictly dominated alpha vectors (comparing vectors only, not at specific belief points). 

The original `intersection_distance` implementation did not return the actual distance from the alpha vectors intersection to the belief point. The original implementation took advantage of the sparsity of the belief vector and only calculated the denominator (the squared difference of the alpha vectors) for non-zero values of the belief vector. This was a good idea for speed, but I think it was an error. For example, check what the original function would return if you had `[1.0, 0.0]` for `a1`, `[0.0, 1.0]` for `a2`, and `[0.5, 0.5]` for `b`. The function would return `1.0` as the distance, but the actual distance is `0.707`. This PR has a revised implementation. It is a bit slower but returns a more accurate distance.

Based on the slightly slower distance calculation, I added `prune_strictly_dominated!(::SARSOPTree)` to prune strictly dominated alpha vectors. This function removes any vectors that are dominated across the entire belief space (by just comparing the values of the vectors, not at specific belief points). This function is called in `prune!(::SARSOPSolver, ::SARSOPTree)` after the beliefs are pruned. This addition ended up speeding up the process a bit and almost made up for the slower distance calculation (still saw some slower times in the larger problems, e.g. RockSample(15,10)).

The last change is to `prune_alpha!`. Instead of looping through each vector and then checking at all belief points, we now first check for the dominating alpha vectors at each belief point. Then we only need to compare the dominated alpha vector at each point to all other vectors to check for delta dominance. This change doesn't catch duplicate alpha vectors, but the added `prune_strictly_dominated!` function catches those. This method also sets up the framework towards an adaptive delta implementation ([related issue](https://github.com/JuliaPOMDP/NativeSARSOP.jl/issues/2)) as we need to track which belief points each vector is dominant at.

I did some ablation studies for each change. However, I have only included the final performance comparisons here for different delta values. The only hit in performance is RockSample(5,5), but the larger problems see a decent performance increase.


# Performance Comparison (delta = 0.1)
## BabyPOMDP
Settings:
  - epsilon: 0.1
  - precision: 0.001
  - delta: 0.1
  - max_steps: 50 (for benchmarking)
  - max_time: 5.0 (for policy run)
### Benchmark
#### Original
```julia
BenchmarkTools.Trial: 3171 samples with 1 evaluation.
 Range (min … max):  1.505 ms …  43.395 ms  ┊ GC (min … max): 0.00% … 96.20%
 Time  (median):     1.537 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.575 ms ± 773.872 μs  ┊ GC (mean ± σ):  1.94% ±  5.34%

    ▂▇██▆▆▃▅▅▃▂▁                                               
  ▃▆████████████▇▆▆▆▅▄▅▄▅▃▄▃▃▃▃▃▃▃▂▂▂▃▂▃▂▂▂▂▁▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂ ▄
  1.51 ms         Histogram: frequency by time         1.7 ms <

 Memory estimate: 425.98 KiB, allocs estimate: 3847.
```
#### New
```julia
BenchmarkTools.Trial: 3639 samples with 1 evaluation.
 Range (min … max):  1.296 ms …   7.054 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.341 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.373 ms ± 289.663 μs  ┊ GC (mean ± σ):  1.46% ± 5.73%

       ▆█▅▅▂                                                   
  ▃▄▅▆███████▇▅▄▄▄▃▃▃▃▂▂▂▂▂▂▂▂▂▁▁▂▂▂▁▁▁▂▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▃
  1.3 ms          Histogram: frequency by time        1.65 ms <

 Memory estimate: 367.88 KiB, allocs estimate: 3862.
```
### Policy Run
#### Original
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.00       0          -28.3501795  -15.6037314  12.7464480985   3          30        
 0.00       10         -16.3057342  -16.2819897  0.0237444537    2          98        
 0.00       20         -16.3054833  -16.3024060  0.0030772894    2          133       
--------------------------------------------------------------------------------------
 0.00       28         -16.3054833  -16.3045949  0.0008883805    2          134       
--------------------------------------------------------------------------------------
```
#### New
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.00       0          -28.3501795  -15.6037314  12.7464480985   3          30        
 0.00       10         -16.3057342  -16.2819897  0.0237444537    2          98        
 0.00       20         -16.3054833  -16.3024060  0.0030772894    2          133       
--------------------------------------------------------------------------------------
 0.00       28         -16.3054833  -16.3045949  0.0008883805    2          134       
--------------------------------------------------------------------------------------
```
## TigerPOMDP
Settings:
  - epsilon: 0.1
  - precision: 0.001
  - delta: 0.1
  - max_steps: 50 (for benchmarking)
  - max_time: 5.0 (for policy run)
### Benchmark
#### Original
```julia
BenchmarkTools.Trial: 175 samples with 1 evaluation.
 Range (min … max):  27.593 ms … 68.988 ms  ┊ GC (min … max): 0.00% … 59.07%
 Time  (median):     28.125 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   28.701 ms ±  3.273 ms  ┊ GC (mean ± σ):  1.37% ±  4.96%

   ▂ ▃▄▂█▂▁▁                                                   
  ▄█████████▅▅▅▄▄▄▃▃▁▁▃▁▄▅▃▃▃▁▁▁▃▁▁▁▁▃▁▁▃▃▁▁▁▁▁▁▁▃▁▁▃▃▁▃▁▁▃▁▃ ▃
  27.6 ms         Histogram: frequency by time        32.1 ms <

 Memory estimate: 3.99 MiB, allocs estimate: 30581.
```
#### New
```julia
BenchmarkTools.Trial: 186 samples with 1 evaluation.
 Range (min … max):  25.873 ms … 42.620 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     26.236 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.947 ms ±  2.222 ms  ┊ GC (mean ± σ):  1.49% ± 4.60%

  ▆█▆▃▁▁ ▃▁                                                    
  █████████▇▁▁▁▁▁▁▁▁█▁▁▄▁▁▄▄▄▁▁▁▁▁▁▄▁▄▁▁▄▁▁▁▁▁▁▁▁▁▆▁▁▄▁▁▁▁▁▁▄ ▄
  25.9 ms      Histogram: log(frequency) by time      36.1 ms <

 Memory estimate: 3.42 MiB, allocs estimate: 30605.
```
### Policy Run
#### Original
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.00       0          -10.7779872  87.0980496   97.8760368879   7          44        
 0.01       10         14.2589622   51.5049954   37.2460332328   5          464       
 0.01       20         18.6214722   33.7498003   15.1283280569   5          649       
 0.02       30         19.3534084   21.8312080   2.4777995367    5          657       
 0.02       40         19.3709835   19.6684906   0.2975071211    5          463       
 0.03       50         19.3713674   19.3833253   0.0119579046    6          212       
--------------------------------------------------------------------------------------
 0.04       57         19.3713684   19.3722266   0.0008581957    6          467       
--------------------------------------------------------------------------------------
```
#### New
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.00       0          -10.7779872  87.0980496   97.8760368879   7          44        
 0.01       10         14.2589622   51.5049954   37.2460332328   5          464       
 0.01       20         18.6214722   33.7498003   15.1283280569   5          649       
 0.02       30         19.3534084   21.8312080   2.4777995367    5          657       
 0.02       40         19.3709835   19.6684906   0.2975071211    5          463       
 0.03       50         19.3713674   19.3833253   0.0119579046    5          212       
--------------------------------------------------------------------------------------
 0.03       57         19.3713684   19.3722266   0.0008581957    5          467       
--------------------------------------------------------------------------------------
```
## RockSamplePOMDP(5,5)
Settings:
  - epsilon: 0.1
  - precision: 0.001
  - delta: 0.1
  - max_steps: 50 (for benchmarking)
  - max_time: 5.0 (for policy run)
### Benchmark
#### Original
```julia
BenchmarkTools.Trial: 274 samples with 1 evaluation.
 Range (min … max):  17.806 ms …  21.571 ms  ┊ GC (min … max): 0.00% … 14.15%
 Time  (median):     18.020 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   18.259 ms ± 717.625 μs  ┊ GC (mean ± σ):  1.14% ±  3.20%

   ▄██▇▆▅▂                                                      
  ▆█████████▄▆▁▁▆▁▄▁▁▁▁▁▁▁▁▁▁▄▁▁▁▁▁▄▄▁▄▁▄▇▄▁▁▁▁▄▁▇▁▄▁▁▄▄█▄▄▄▄▄ ▆
  17.8 ms       Histogram: log(frequency) by time      20.8 ms <

 Memory estimate: 4.50 MiB, allocs estimate: 27979.
```
#### New
```julia
BenchmarkTools.Trial: 254 samples with 1 evaluation.
 Range (min … max):  18.852 ms … 37.836 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     19.165 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   19.686 ms ±  1.788 ms  ┊ GC (mean ± σ):  2.03% ± 5.28%

   ▆█▇▃▁                                                       
  ███████▄▁▄▄▄▁▄▁▁▁▄▁▁▄▅▁▁▅▁▁▁▁▁▅▁▁▄▄▁▅▄▁▁▁▁▁▄▄▁▆▁▅▄▁▁▁▁▁▄▅▁▄ ▅
  18.9 ms      Histogram: log(frequency) by time      25.1 ms <

 Memory estimate: 4.49 MiB, allocs estimate: 28244.
```
### Policy Run
#### Original
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.00       0          15.2423183   18.3402991   3.0979807840    13         14        
 0.00       10         16.9264164   18.1564052   1.2299888405    50         85        
 0.01       20         16.9264164   17.6433978   0.7169814681    68         113       
 0.01       30         16.9264164   17.5412846   0.6148682461    75         138       
 ...      
 0.10       230        16.9264164   16.9559859   0.0295695736    125        334       
 0.11       240        16.9264164   16.9499615   0.0235451701    132        333       
 0.11       250        16.9264164   16.9352883   0.0088718960    135        212       
 0.11       260        16.9264164   16.9318942   0.0054778577    125        95        
--------------------------------------------------------------------------------------
 0.12       269        16.9264164   16.9264164   -0.0000000000   127        42        
--------------------------------------------------------------------------------------
```
#### New
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.00       0          15.2423183   18.3402991   3.0979807840    13         14        
 0.00       10         16.9264164   18.1564052   1.2299888405    50         85        
 0.01       20         16.9264164   17.6433978   0.7169814681    71         113       
 0.01       30         16.9264164   17.5412846   0.6148682461    82         138       
 ...    
 0.14       250        16.9264164   16.9439586   0.0175422648    140        377       
 0.14       260        16.9264164   16.9341914   0.0077750094    140        294       
 0.15       270        16.9264164   16.9318942   0.0054778577    140        270       
 0.15       280        16.9264164   16.9291465   0.0027301069    140        236       
--------------------------------------------------------------------------------------
 0.16       290        16.9264164   16.9273809   0.0009645004    140        103       
--------------------------------------------------------------------------------------
```
## TagPOMDP
Settings:
  - epsilon: 0.1
  - precision: 0.001
  - delta: 0.1
  - max_steps: 50 (for benchmarking)
  - max_time: 5.0 (for policy run)
### Benchmark
#### Original
```julia
BenchmarkTools.Trial: 3 samples with 1 evaluation.
 Range (min … max):  1.842 s …   1.889 s  ┊ GC (min … max): 0.00% … 2.55%
 Time  (median):     1.865 s              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.865 s ± 23.570 ms  ┊ GC (mean ± σ):  0.86% ± 1.47%

  █                          █                            █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  1.84 s         Histogram: frequency by time        1.89 s <

 Memory estimate: 149.51 MiB, allocs estimate: 838113.
```
#### New
```julia
BenchmarkTools.Trial: 6 samples with 1 evaluation.
 Range (min … max):  903.588 ms … 924.269 ms  ┊ GC (min … max): 4.93% … 4.11%
 Time  (median):     907.070 ms               ┊ GC (median):    4.43%
 Time  (mean ± σ):   911.095 ms ±   8.801 ms  ┊ GC (mean ± σ):  4.58% ± 0.58%

  █ █      ██                                     █           █  
  █▁█▁▁▁▁▁▁██▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁█ ▁
  904 ms           Histogram: frequency by time          924 ms <

 Memory estimate: 144.89 MiB, allocs estimate: 838603.
```
### Policy Run
#### Original
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.01       0          -19.3713764  -4.6944443   14.6769321680   18         48        
 0.14       10         -12.1605923  -4.9155579   7.2450343269    192        345       
 0.40       20         -11.9445215  -5.0085441   6.9359773437    285        545       
 0.77       30         -11.5468985  -5.0615592   6.4853393203    361        773       
 ...     
 2.94       70         -11.2132678  -5.2020310   6.0112367924    617        1446      
 3.68       80         -11.1744215  -5.2321641   5.9422574180    632        1578      
 4.08       90         -11.1744215  -5.2645196   5.9099018818    703        1675      
 4.99       100        -11.1403313  -5.2904604   5.8498708373    708        1833      
--------------------------------------------------------------------------------------
 5.01       102        -11.1403313  -5.2905198   5.8498114519    722        1843      
--------------------------------------------------------------------------------------
```
#### New
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.01       0          -19.3713764  -4.6944443   14.6769321680   18         48        
 0.09       10         -12.1605923  -4.9155579   7.2450343269    108        345       
 0.19       20         -11.9445215  -5.0085441   6.9359773437    162        545       
 0.33       30         -11.5468985  -5.0615592   6.4853393203    210        774       
 ...   
 3.44       130        -11.1282380  -5.3416163   5.7866217372    544        2250      
 3.93       140        -11.1202809  -5.3649098   5.7553711760    575        2380      
 4.38       150        -11.1016246  -5.3819350   5.7196895946    608        2516      
 4.85       160        -11.1016246  -5.3932110   5.7084135463    612        2633      
--------------------------------------------------------------------------------------
 5.01       164        -11.1016246  -5.3989383   5.7026863223    613        2675      
--------------------------------------------------------------------------------------
```
## TagPOMDP
Settings:
  - epsilon: 0.1
  - precision: 0.001
  - delta: 0.1
  - max_steps: 50 (for benchmarking)
  - max_time: 60.0 (for policy run)
### Policy Run
#### Original
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.01       0          -19.3713764  -4.6944443   14.6769321680   18         48        
 0.14       10         -12.1605923  -4.9155579   7.2450343269    192        345       
 0.41       20         -11.9445215  -5.0085441   6.9359773437    285        545       
 0.78       30         -11.5468985  -5.0615592   6.4853393203    361        773       
 ...    
 50.40      400        -10.9287465  -5.6406050   5.2881415754    1413       5473      
 53.02      410        -10.9157919  -5.6579474   5.2578445942    1464       5605      
 55.63      420        -10.9157919  -5.6669246   5.2488673921    1469       5707      
 58.36      430        -10.9049055  -5.6747273   5.2301781393    1486       5820      
--------------------------------------------------------------------------------------
 60.85      440        -10.9049055  -5.6831865   5.2217189894    1464       5887      
--------------------------------------------------------------------------------------
```
#### New
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
--------------------------------------------------------------------------------------
 0.01       0          -19.3713764  -4.6944443   14.6769321680   18         48        
 0.09       10         -12.1605923  -4.9155579   7.2450343269    108        345       
 0.19       20         -11.9445215  -5.0085441   6.9359773437    162        545       
 0.33       30         -11.5468985  -5.0615592   6.4853393203    210        774       
 ...    
 52.15      540        -10.9280846  -5.7686969   5.1593877438    1084       7060      
 54.88      550        -10.9227351  -5.7750679   5.1476671438    1100       7249      
 57.15      560        -10.9227351  -5.7818944   5.1408406919    1113       7351      
 59.52      570        -10.9227351  -5.7894943   5.1332407734    1121       7463      
--------------------------------------------------------------------------------------
 60.04      573        -10.9227351  -5.7910739   5.1316611547    1124       7483      
--------------------------------------------------------------------------------------
```
## RockSamplePOMDP(15,10)
Settings:
  - epsilon: 0.1
  - precision: 0.001
  - delta: 0.1
  - max_steps: 50 (for benchmarking)
  - max_time: 120.0 (for policy run)
  - init_lower: BlindLowerBound(9223372036854775807, 60.0, 0.001, Float64[], Float64[])
  - init_upper: FastInformedBound(9223372036854775807, 60.0, 0.001, 0.0, Float64[], Float64[])
### Policy Run
#### Original
```julia
--------------------------------------------------------------------------------------
 Time       Iter       LB           UB           Precision       # Alphas   # Beliefs 
…