@WhiffleFish 2024-07-09
Thanks for the PR! You're entirely correct, the current intersection distance implementation is wrong. However, I think the significant increase in allocations here needs to be addressed.