@dylan-asmar 2024-07-22
@WhiffleFish I updated the benchmark results after the suggested changes. We are doing WAY better on allocations than before and are better than or on par with the original implementation now. 

I left the conversations open on the change requests. We can resolve the conversation if you are satisfied with the comments/changes.

Let me know if there are any other changes we should consider.