@WhiffleFish 2025-03-04
There were a couple of things I still wanted to optimize further here but never got around to finishing, so I think this is definitely good to merge now. 

Thanks for the hard work here, Dylan.