@dylan-asmar 2025-03-04
Thoughts on writing them up in an issue? Just a place to track your ideas and potentially implement the performance improvements when we have time?